fix(aggregator options): moved collation from sortOptions to generic options, since in mongo collation is also available on every aggregator (and maybe we will need that later in another aggregator than sort as well)
