+++ src/aggregator.ts
